requirements.txt
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
aiohttp==3.10.10
peewee==3.17.7
pytest==8.3.3
